+++ CHANGELOG.md
@@ -4,11 +4,10 @@ Changelog
 3.0.0 (unreleased)
 ------------------
 
+* **2020-01-08**: [BC-BREAK] Move feature packages to require-dev to make sure packages using the testing component declare their dependencies correctly.
 * **2019-12-05**: [BC-BREAK] Support PHPUnit 6, 7 and 8, drop support for PHPUnit 5.7
 * **2019-07-26**: [BC-BREAK] Remove deprecated `BaseTestCase::getClient`, use `BaseTestCase::getFrameworkBundleClient` instead.
 * **2019-02-28**: [BC-BREAK] Remove `DatabaseTestListener`, use scripts in `bin/make/` to ramp up your testing environment for several test suites
-* **2019-02-28**: [BC-BREAK] Introduce PHPUnit 6 Support, Remove PHPUnit 5.7.
-* **2019-02-28**: Add testing scenarios for Symfony 4.2 and 4.3 (allowed to fail for now)
 
 2.1.12
 ------
@@ -19,7 +18,7 @@ Changelog
 2.1.0
 -----
 
-* **2017-01-18**: Introduce a `TestCompilePass` to make services public, which are needed 
+* **2017-01-18**: Introduce a `TestCompilePass` to make services public, which are needed
 in i.e. a WebTest.
  * **2017-11-08**: Removed php 5.6 and 7.0 support, removed Symfony 3.0.* and 3.1.* support
  introduce KERNEL_CLASS handling to avoid deprecated KERNEL_DIR, removed usage of `ProcessBuilder`